# pkg/container-collection: replace container iteration with lookup maps for hot paths