When enriching container information on events using the container-collection (which is a hot path in our code), we usually look up containers by either mntnsid or netnsid. Previously, we returned matching containers after iterating over the main container map. With this patch, we instead use dedicated lookup maps for both mntns and netns to increase performance. Additionally, previously, when looking up containers using netnsid, a temporary array got created - this allocation has also been moved to the map maintenance and only is done when adding or removing containers instead of on each lookup operation.

Benchmarks:

```
$ benchstat main.bench update.bench
goos: linux
goarch: arm64
pkg: github.com/inspektor-gadget/inspektor-gadget/pkg/container-collection
                         │   main.bench    │            update.bench             │
                         │     sec/op      │   sec/op     vs base                │
LookupContainerByMntns-4    52008.00n ± 1%   42.48n ± 0%  -99.92% (p=0.000 n=10)
LookupContainerByNetns-4   104945.00n ± 1%   41.47n ± 1%  -99.96% (p=0.000 n=10)
geomean                        73.88µ        41.98n       -99.94%
```