@flyth 2023-07-03
> Good optimisation, I just would like to have the benchmark result in the commit message to get a better understanding of the gain.

Added.

Thanks for the reviews!